[envsec] Move envcli to envsec/pkg
