+++ envsec/pkg/envcli/rm.go
